Tests/LibWeb/Text/expected/hit_testing/wrapped-span-inter-line-gap.txt

rects: 2
gap hit is container: true
gap hit is span or its text: false
line 2 hit is span text: true
